third_party/WebKit/LayoutTests/accessibility/input-type-password-value-and-selection-expected.txt

This tests that AXValue, caret position and text selection are reported correctly for password fields.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 0
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.stringValue is "AXValue: ••••••"
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 0
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 1
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 2
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 3
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 4
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 1
PASS axPassword.selectionEnd is 1
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 1
PASS axPassword.selectionEnd is 2
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 1
PASS axPassword.selectionEnd is 3
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 1
PASS axPassword.selectionEnd is 4
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 1
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 2
PASS axPassword.selectionEnd is 2
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 2
PASS axPassword.selectionEnd is 3
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 2
PASS axPassword.selectionEnd is 4
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 2
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 3
PASS axPassword.selectionEnd is 3
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 3
PASS axPassword.selectionEnd is 4
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 3
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 4
PASS axPassword.selectionEnd is 4
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 4
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 5
PASS axPassword.selectionEnd is 5
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS axPassword.selectionStart is 0
PASS axPassword.selectionEnd is 0
PASS axPassword.selectionStartLineNumber is 0
PASS axPassword.selectionEndLineNumber is 0
PASS successfullyParsed is true

TEST COMPLETE
